Share extension hosts the real agent builder; Make commits via shared AgentCreationFlow
Targetless shares now present the actual AgentDraftComposer instead of
improvised UI, and tapping Make finishes the agent from the extension
itself - no handoff into the app.

- Extract AgentDraftComposer into the ConvosComposer package,
  genericized over an AgentDraftComposing protocol. Connections become
  a generic chip slot (AgentDraftConnectionChip) the app fills;
  voice-memo and video support are capability flags the extension
  turns off. VoiceMemoAttachmentChip moves along with it, and
  DesignConstants gains public textPrimary/textSecondary tokens.
- Extract the Make commit recipe into AgentCreationFlow in ConvosCore:
  paired attachment compression + creation-card chips, card thumbnail
  generation, prompt message-id allocation, generation submission +
  creation-prompt card persistence, and the ready/slug waits. Both
  AgentBuilderViewModel and the extension's makeAgent now commit
  through it, so the two hosts cannot drift.
- The extension's Make creates the conversation, submits the
  generation (photos ride the generation API), persists the creation
  card so the shared image is visible in the new chat, and publishes
  the prompt under the card's bundled message id with the publish
  runway held.
- Remove the AgentBuilderShareHandoff record and its app-side
  reopen path; the builder no longer reappears in the app after a
  share-extension Make.
